refactor: Max 원시값을 포장하기
